@@ -38,19 +38,19 @@ Rails.application.configure do
 
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true
-  config.action_mailer.default_url_options = { :host => 'localhost:3000' }
+  config.action_mailer.default_url_options = { host: 'localhost:3000' }
   # config.action_mailer.delivery_method = :smtp
   # config.action_mailer.smtp_settings = {:address => "localhost", :port => 1025}
-  
+
   config.action_mailer.raise_delivery_errors = true
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.perform_deliveries = true
   ActionMailer::Base.smtp_settings = {
-     :address              => "smtp.gmail.com",
-     :port                 => 587,
-     :user_name            => Rails.application.secrets.gmail_username,
-     :password             => Rails.application.secrets.gmail_password,
-     :authentication       => "plain",
-     :enable_starttls_auto => true
-   }
+    address: 'smtp.gmail.com',
+    port: 587,
+    user_name: Rails.application.secrets.gmail_username,
+    password: Rails.application.secrets.gmail_password,
+    authentication: 'plain',
+    enable_starttls_auto: true
+  }
 end
