@@ -37,6 +37,8 @@ Rails.application.configure do
 
   # `config.assets.precompile` and `config.assets.version` have moved to config/initializers/assets.rb
 
+  config.force_ssl = true
+
   # Specifies the header that your server uses for sending files.
   # config.action_dispatch.x_sendfile_header = 'X-Sendfile' # for Apache
   # config.action_dispatch.x_sendfile_header = 'X-Accel-Redirect' # for NGINX
@@ -81,11 +83,11 @@ Rails.application.configure do
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.perform_deliveries = true
   ActionMailer::Base.smtp_settings = {
-     :address              => "smtp.gmail.com",
-     :port                 => 587,
-     :user_name            => Rails.application.secrets.gmail_username,
-     :password             => Rails.application.secrets.gmail_password,
-     :authentication       => "plain",
-     :enable_starttls_auto => true
-   }
+    address: 'smtp.gmail.com',
+    port: 587,
+    user_name: Rails.application.secrets.gmail_username,
+    password: Rails.application.secrets.gmail_password,
+    authentication: 'plain',
+    enable_starttls_auto: true
+  }
 end
