@@ -1,9 +1,8 @@
 require 'test_helper'
 
 class HomeControllerTest < ActionController::TestCase
-  test "should get index" do
+  test 'should get index' do
     get :index
     assert_response :success
   end
-
 end
