@@ -1,4 +1,7 @@
-ENV["RAILS_ENV"] = "test"
+require 'coveralls'
+Coveralls.wear!
+
+ENV['RAILS_ENV'] = 'test'
 require File.expand_path('../../config/environment', __FILE__)
 require 'rails/test_help'
 
@@ -10,4 +13,12 @@ class ActiveSupport::TestCase
   fixtures :all
 
   # Add more helper methods to be used by all tests here...
+end
+
+class ActionController::TestCase
+  include Devise::TestHelpers
+end
+
+class ActionView::TestCase
+  include Devise::TestHelpers
 end
